+++ plugin/backend.go
@@ -29,20 +29,12 @@ type ServiceBackend interface {
 	// IsNameError return true if err indicated a record not found condition
 	IsNameError(err error) bool
 
-	Transferer
-}
-
-// Transferer defines an interface for backends that provide AXFR of all records.
-type Transferer interface {
 	// Serial returns a SOA serial number to construct a SOA record.
 	Serial(state request.Request) uint32
 
 	// MinTTL returns the minimum TTL to be used in the SOA record.
 	MinTTL(state request.Request) uint32
 
-	// Transfer handles a zone transfer it writes to the client just
-	// like any other handler.
-	Transfer(ctx context.Context, state request.Request) (int, error)
 }
 
 // Options are extra options that can be specified for a lookup.
