+++ plugin/kubernetes/xfr_test.go
@@ -5,14 +5,13 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/coredns/coredns/plugin/kubernetes/object"
 	"github.com/coredns/coredns/plugin/pkg/dnstest"
 	"github.com/coredns/coredns/plugin/test"
 
 	"github.com/miekg/dns"
 )
 
-func TestKubernetesXFR(t *testing.T) {
+func TestKubernetesTransferDeprecated(t *testing.T) {
 	k := New([]string{"cluster.local."})
 	k.APIConn = &APIConnServeTest{}
 	k.TransferTo = []string{"10.240.0.1:53"}
@@ -124,106 +123,5 @@ func TestKubernetesXFRNotAllowed(t *testing.T) {
 	if len(w.Msgs[0].Answer) != 0 {
 		t.Logf("%+v\n", w)
 		t.Fatal("Got an answer, should not have")
-	}
-}
-
-// difference shows what we're missing when comparing two RR slices
-func difference(testRRs []dns.RR, gotRRs []dns.RR) []dns.RR {
-	expectedRRs := map[string]struct{}{}
-	for _, rr := range testRRs {
-		expectedRRs[rr.String()] = struct{}{}
-	}
-
-	foundRRs := []dns.RR{}
-	for _, rr := range gotRRs {
-		if _, ok := expectedRRs[rr.String()]; !ok {
-			foundRRs = append(foundRRs, rr)
-		}
-	}
-	return foundRRs
-}
-
-func TestEndpointsEquivalent(t *testing.T) {
-	epA := object.Endpoints{
-		Subsets: []object.EndpointSubset{{
-			Addresses: []object.EndpointAddress{{IP: "1.2.3.4", Hostname: "foo"}},
-		}},
-	}
-	epB := object.Endpoints{
-		Subsets: []object.EndpointSubset{{
-			Addresses: []object.EndpointAddress{{IP: "1.2.3.4", Hostname: "foo"}},
-		}},
-	}
-	epC := object.Endpoints{
-		Subsets: []object.EndpointSubset{{
-			Addresses: []object.EndpointAddress{{IP: "1.2.3.5", Hostname: "foo"}},
-		}},
-	}
-	epD := object.Endpoints{
-		Subsets: []object.EndpointSubset{{
-			Addresses: []object.EndpointAddress{{IP: "1.2.3.5", Hostname: "foo"}},
-		},
-			{
-				Addresses: []object.EndpointAddress{{IP: "1.2.2.2", Hostname: "foofoo"}},
-			}},
-	}
-	epE := object.Endpoints{
-		Subsets: []object.EndpointSubset{{
-			Addresses: []object.EndpointAddress{{IP: "1.2.3.5", Hostname: "foo"}, {IP: "1.1.1.1"}},
-		}},
-	}
-	epF := object.Endpoints{
-		Subsets: []object.EndpointSubset{{
-			Addresses: []object.EndpointAddress{{IP: "1.2.3.4", Hostname: "foofoo"}},
-		}},
-	}
-	epG := object.Endpoints{
-		Subsets: []object.EndpointSubset{{
-			Addresses: []object.EndpointAddress{{IP: "1.2.3.4", Hostname: "foo"}},
-			Ports:     []object.EndpointPort{{Name: "http", Port: 80, Protocol: "TCP"}},
-		}},
-	}
-	epH := object.Endpoints{
-		Subsets: []object.EndpointSubset{{
-			Addresses: []object.EndpointAddress{{IP: "1.2.3.4", Hostname: "foo"}},
-			Ports:     []object.EndpointPort{{Name: "newportname", Port: 80, Protocol: "TCP"}},
-		}},
-	}
-	epI := object.Endpoints{
-		Subsets: []object.EndpointSubset{{
-			Addresses: []object.EndpointAddress{{IP: "1.2.3.4", Hostname: "foo"}},
-			Ports:     []object.EndpointPort{{Name: "http", Port: 8080, Protocol: "TCP"}},
-		}},
-	}
-	epJ := object.Endpoints{
-		Subsets: []object.EndpointSubset{{
-			Addresses: []object.EndpointAddress{{IP: "1.2.3.4", Hostname: "foo"}},
-			Ports:     []object.EndpointPort{{Name: "http", Port: 80, Protocol: "UDP"}},
-		}},
-	}
-
-	tests := []struct {
-		equiv bool
-		a     *object.Endpoints
-		b     *object.Endpoints
-	}{
-		{true, &epA, &epB},
-		{false, &epA, &epC},
-		{false, &epA, &epD},
-		{false, &epA, &epE},
-		{false, &epA, &epF},
-		{false, &epF, &epG},
-		{false, &epG, &epH},
-		{false, &epG, &epI},
-		{false, &epG, &epJ},
-	}
-
-	for i, tc := range tests {
-		if tc.equiv && !endpointsEquivalent(tc.a, tc.b) {
-			t.Errorf("Test %d: expected endpoints to be equivalent and they are not.", i)
-		}
-		if !tc.equiv && endpointsEquivalent(tc.a, tc.b) {
-			t.Errorf("Test %d: expected endpoints to be seen as different but they were not.", i)
-		}
 	}
 }
